LayoutTests/js/string-slice-abnormal-values-expected.txt

This test checks for handling of abnormal values passed to String.slice

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "abc".slice(0) is "abc"
PASS "abc".slice(0, Infinity) is "abc"
PASS successfullyParsed is true

TEST COMPLETE
